Add support for generic class generation
Motivation
----
Rather than returning mixed, you can return an expected type with a
generic function if you know the type ahead of time. This lets you
coerrce a generated class in functions that require specific types. This
is more refined than 'mixed' so it allows you to fix the typechecker
errors.