Add support for importing vector collectable
Motivation
----
We previously did not have a way of importing a storable object given a
collection. With this change, we introduce a type safe way of importing
a type using dynamic class creation along with type hints to make allow
us to import a list of storable objects. If the type hinted at does not
match the collection constraint, the Hacklang runtime will generated an
exception. See: https://docs.hhvm.com/hack/generics/erasure

How do you use this?
----
Checkout the unit tests for usage.

Why was the previous collection not just updated?
---
The previous collection made no assumptions for the refined type a
collection wrapped. Because of this, you could stand up a storable
vector of non storable objects and would not see an error when an import
was attempted. In addition, type erasure for the language made it
impossible for us to stand up the expected type when we tried to
deserialize a payload.

Given these issues, a new version of the collection was created. The old
version of the collection has since been deprecated for this version. In
order to not break current use cases, we will not delete the old
implementation for now.