Add library CodePath support and change inclusions of XHProfiler
Motivation
----
Currently using XHProfiler meant that the code path needed to have the
library found in the top level root path of whatever you had set as your
application root. Because we want to support library paths that may be
unique per project, we now add support of specifying paths for different
libraries as part of the CodePath environment.

In addition, XHProfiler has been modified to account for this support.