Add support for locking memcache
Motivation
---
We would like to be able to use Memcache as a simple distributed locking
system. In order to do so, we have added support for locking in the
Cache API. This first checkin is to to get the necessary skeleton in to
get this started.

How this will be supported is through Memcache clusters that will host
the lock keys with policies to never evict. That will be a separate
discussion at a later date.